+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/components/content/settings/main.js
@@ -8,11 +8,14 @@
  */
 
 define([
+    'underscore',
+    'jquery',
     'config',
     'sulusecurity/components/users/models/user',
+    'sulucontact/models/contact',
     'services/husky/url-validator',
     'sulucontent/services/content-manager'
-], function(Config, User, urlValidator, contentManager) {
+], function(_, $, Config, User, Contact, urlValidator, contentManager) {
 
     'use strict';
 
@@ -45,6 +48,8 @@ define([
                 linkContainer: '#external-link-container .link'
             }
         },
+
+        authorFullname = null,
 
         isShadow = function() {
             return this.sandbox.dom.prop('#shadow_on_checkbox', 'checked');
@@ -94,6 +99,32 @@ define([
             }
 
             this.sandbox.dom.text('#changed', changedText);
+        },
+
+        setAuthorChangelog = function(fullName, time) {
+            var authoredText, formattedTime = this.sandbox.date.format(time);
+
+            fullName = fullName || authorFullname;
+
+            if (!!fullName) {
+                authorFullname = fullName;
+                authoredText = this.sandbox.util.sprintf(
+                    this.sandbox.translate('sulu.content.form.settings.changelog.authored'),
+                    {
+                        author: fullName,
+                        authored: formattedTime
+                    }
+                );
+            } else {
+                authoredText = this.sandbox.util.sprintf(
+                    this.sandbox.translate('sulu.content.form.settings.changelog.authored-only'),
+                    {
+                        authored: formattedTime
+                    }
+                )
+            }
+
+            this.sandbox.dom.text('#author', authoredText);
         },
 
         showChangelogContainer = function() {
@@ -230,6 +261,9 @@ define([
                 this.updateVisibilityForShadowCheckbox(false);
             }.bind(this));
 
+            this.sandbox.dom.on('#change-author', 'click', function() {
+                this.openAuthorSelection();
+            }.bind(this));
         },
 
         updateVisibilityForShadowCheckbox: function(isInitial) {
@@ -366,8 +400,10 @@ define([
         updateChangelog: function(data) {
             var creator,
                 changer,
+                author,
                 creatorDef = this.sandbox.data.deferred(),
-                changerDef = this.sandbox.data.deferred();
+                changerDef = this.sandbox.data.deferred(),
+                authorDef = this.sandbox.data.deferred();
 
             if (data.creator === data.changer) {
                 creator = new User({id: data.creator});
@@ -408,9 +444,27 @@ define([
                 })
             }
 
-            this.sandbox.data.when(creatorDef, changerDef).then(function(creation, change) {
+            if (!!data.author) {
+                author = new Contact({id: data.author});
+                author.fetch({
+                    global: false,
+
+                    success: function(model) {
+                        authorDef.resolve(model.get('fullName'), new Date(data.authored));
+                    }.bind(this),
+
+                    error: function() {
+                        authorDef.resolve(null, new Date(data.authored));
+                    }.bind(this)
+                });
+            } else {
+                authorDef.resolve(null, new Date(data.authored));
+            }
+
+            this.sandbox.data.when(creatorDef, changerDef, authorDef).then(function(creation, change, author) {
                 setCreationChangelog.call(this, creation[0], creation[1]);
                 setChangeChangelog.call(this, change[0], change[1]);
+                setAuthorChangelog.call(this, author[0], author[1]);
                 showChangelogContainer.call(this);
             }.bind(this));
         },
@@ -473,11 +527,10 @@ define([
         submit: function(action) {
             this.sandbox.logger.log('save Model');
 
-            var data = {
-                    id: this.data.id
-                },
+            var data = {},
                 baseLanguages = this.sandbox.dom.data('#shadow_base_language_select', 'selectionValues');
 
+            data.id = this.data.id;
             data.navContexts = this.sandbox.dom.data('#nav-contexts', 'selection');
             data.nodeType = parseInt(this.sandbox.dom.val('input[name="nodeType"]:checked'));
             data.shadowOn = isShadow.call(this);
@@ -500,6 +553,9 @@ define([
             if (!!data.shadowOn && !!baseLanguages && baseLanguages.length > 0) {
                 data.shadowBaseLanguage = baseLanguages[0];
             }
+
+            data.author = this.data.author;
+            data.authored = this.data.authored;
 
             if (!this.validate(data)) {
                 this.sandbox.emit('sulu.labels.warning.show', 'form.validation-warning', 'labels.warning');
@@ -616,6 +672,65 @@ define([
             }
 
             return result;
+        },
+
+        openAuthorSelection: function() {
+            var $overlayContainer = $('<div/>'),
+                $componentContainer = $('<div/>');
+
+            this.$el.append($overlayContainer);
+
+            this.sandbox.start([{
+                name: 'overlay@husky',
+                options: {
+                    el: $overlayContainer,
+                    instanceName: 'author-selection',
+                    openOnStart: true,
+                    removeOnClose: true,
+                    skin: 'medium',
+                    slides: [
+                        {
+                            title: this.sandbox.translate('sulu.content.form.settings.author'),
+                            okCallback: function() {
+                                this.sandbox.emit('sulu.content.contents.get-author');
+                            }.bind(this),
+                            data: $componentContainer
+                        }
+                    ]
+                }
+            }]);
+
+            this.sandbox.once('husky.overlay.author-selection.initialized', function() {
+                this.sandbox.start([
+                    {
+                        name: 'content/settings/author-selection@sulucontent',
+                        options: {
+                            el: $componentContainer,
+                            locale: this.options.locale,
+                            data: {author: this.data.author, authored: this.data.authored},
+                            selectCallback: function(data) {
+                                this.setAuthor(data);
+
+                                this.sandbox.emit('husky.overlay.author-selection.close');
+                            }.bind(this)
+                        }
+                    }
+                ]);
+            }.bind(this));
+        },
+
+        setAuthor: function(data) {
+            this.setHeaderBar(false);
+
+            this.data.authored = data.authored;
+            if (!data.authorItem) {
+                setAuthorChangelog.call(this, null, new Date(data.authored));
+
+                return;
+            }
+
+            setAuthorChangelog.call(this, data.authorItem.firstName + ' ' + data.authorItem.lastName, new Date(data.authored));
+            this.data.author = data.author;
         }
     };
 });