+++ src/Sulu/Bundle/ContentBundle/Tests/Functional/Controller/NodeControllerTest.php
@@ -136,6 +136,7 @@ class NodeControllerTest extends SuluTestCase
         $this->assertEquals(['tag1', 'tag2'], $response->tags);
         $this->assertEquals($this->getTestUserId(), $response->creator);
         $this->assertEquals($this->getTestUserId(), $response->changer);
+        $this->assertNotFalse(\DateTime::createFromFormat('Y-m-d', $response->authored));
 
         /** @var NodeInterface $content */
         $defaultContent = $this->session->getNode('/cmf/sulu_io/contents/news/test-1');
@@ -263,6 +264,7 @@ class NodeControllerTest extends SuluTestCase
         $this->assertEquals(['tag1', 'tag2'], $response['tags']);
         $this->assertEquals('/test_en', $response['url']);
         $this->assertEquals('Test English', $response['article']);
+        $this->assertNotFalse(\DateTime::createFromFormat('Y-m-d', $response['authored']));
 
         $client->request('GET', '/api/nodes/' . $document->getUuid() . '?language=de');
         $response = json_decode($client->getResponse()->getContent(), true);
@@ -508,8 +510,11 @@ class NodeControllerTest extends SuluTestCase
                 'title' => 'Testtitle DE',
                 'template' => 'default',
                 'url' => '/test-de',
+                'authored' => '2016-12-21',
+                'author' => 1,
             ]
         );
+        $this->assertHttpStatusCode(200, $client->getResponse());
         $client->request(
             'PUT',
             '/api/nodes/' . $response['id'] . '?webspace=sulu_io&language=en',
@@ -517,8 +522,11 @@ class NodeControllerTest extends SuluTestCase
                 'title' => 'Testtitle EN',
                 'template' => 'default',
                 'url' => '/test-en',
+                'authored' => '2016-12-21',
+                'author' => 1,
             ]
         );
+        $this->assertHttpStatusCode(200, $client->getResponse());
 
         $client->request('GET', '/api/nodes/' . $response['id'] . '?language=de');
         $response = json_decode($client->getResponse()->getContent(), true);
@@ -529,6 +537,10 @@ class NodeControllerTest extends SuluTestCase
         $this->assertEquals($this->getTestUserId(), $response['changer']);
         $this->assertEquals($this->getTestUserId(), $response['creator']);
 
+        $this->assertNotFalse(\DateTime::createFromFormat('Y-m-d', $response['authored']));
+        $this->assertEquals('2016-12-21', (new \DateTime($response['authored']))->format('Y-m-d'));
+        $this->assertEquals(1, $response['author']);
+
         $client->request('GET', '/api/nodes/' . $response['id'] . '?language=en');
         $response = json_decode($client->getResponse()->getContent(), true);
 
@@ -537,6 +549,9 @@ class NodeControllerTest extends SuluTestCase
         $this->assertEquals(false, $response['publishedState']);
         $this->assertEquals($this->getTestUserId(), $response['changer']);
         $this->assertEquals($this->getTestUserId(), $response['creator']);
+
+        $this->assertEquals('2016-12-21', (new \DateTime($response['authored']))->format('Y-m-d'));
+        $this->assertEquals(1, $response['author']);
 
         $this->assertFalse(
             $this->liveSession->getNode('/cmf/sulu_io/contents/testtitle-en')->hasProperty('i18n:en-changed')